stopped sound from repeating when key is held
added a keyState object that logs whether a key remains pressed. When a
keydown event is fired. The script first checks if that keycode's
value in the keyState object is set to 'true'. If not, fire keydown event
to play sound and sets its corresponding keyState to 'true'. If its keyState
is already 'true', do nothing. Keyup event will reset the boolean in the
keyState object to 'false'.